+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.36] - 2026-12-08
+### Changed
+- Regular maintenance updates to enhance the integrity and security of the integration
+- Update the Sofinco payment method name on the checkout
+### Fixed
+- Fix Credit card and hosted checkout labels for non-default languages
+
 ## [1.4.35] - 2026-29-07
 ### Fixed
 - Display correct payment method label on capture-confirmation state transition